BE/requirements.txt
gitpython>=3.1.0
graphviz>=0.20.0
requests>=2.31.0
streamlit>=1.28.0
python-dotenv>=1.0.0